[feature] WBJ-259
